+++ CHANGELOG.md
@@ -94,6 +94,7 @@
 - Now, only savanna and desert villagers accept trading bottle gourds and water-filled gourd bottles.
 - Potions, honey bottles, and milk buckets now can be used to fill gourd bottles.
 - Gourd bottle blocks now show the fluid contained graphically
+- Gourd bottle blocks now emit comparator signal
 
 ## Fixes